pyproject.toml
[tool.poetry]
name = "fit-text"
version = "0.1.0"
description = ""
authors = ["Alexey <mr.swasher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
reportlab = "^4.2.0"
borb = "^2.1.22"
fpdf2 = "^2.7.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
